Fix for when video is not in the first card
I didn't really understand why sometimes self.player is different object is different when component is in the first card being loaded in viewer or not (could be api load timing?). And why sometimes a iframe needs to be generated in polymer instead of iframe that is generated by polymer, BUT self.player.d is null! self.player.f.src has the iframe url! (was it a typo or youtube api change?)